+++ src/main/java/net/kccricket/clicksorted/gui/ClickSortedHolder.java
@@ -12,7 +12,7 @@ package net.kccricket.clicksorted.gui;
  * You should have received a copy of the GNU General Public License along with ClickSorted. If not, see <http://www.gnu.org/licenses/>.
  */
 
-import net.kccricket.clicksorted.config.LangConfig;
+import net.kccricket.clicksorted.text.lang.Localized;
 import net.kccricket.clicksorted.text.MessageUtil;
 import org.bukkit.Material;
 import org.bukkit.inventory.InventoryHolder;
@@ -30,7 +30,7 @@ import org.bukkit.inventory.meta.ItemMeta;
 public interface ClickSortedHolder extends InventoryHolder {
 
     /** Black stained-glass pane used as inert filler in all ClickSorted GUIs. */
-    static ItemStack buildFiller(LangConfig lang) {
+    static ItemStack buildFiller(Localized lang) {
         ItemStack pane = new ItemStack(Material.BLACK_STAINED_GLASS_PANE);
         ItemMeta meta = pane.getItemMeta();
         meta.displayName(lang.getColoredMessage("guiFillerName"));
@@ -39,7 +39,7 @@ public interface ClickSortedHolder extends InventoryHolder {
     }
 
     /** Book used as a help/instructions widget in all ClickSorted GUIs. */
-    static ItemStack buildHelpBook(LangConfig lang, String loreKey) {
+    static ItemStack buildHelpBook(Localized lang, String loreKey) {
         ItemStack book = new ItemStack(Material.BOOK);
         ItemMeta meta = book.getItemMeta();
         meta.displayName(lang.getColoredMessage("guiHelpBookName"));